manual support conditional compilation
