+++ .gitignore
@@ -89,3 +89,6 @@ gradlew.*
 # Ignore logs and core dumps
 *.log
 core.*
+
+# Ignore vscode project files
+.vscode/
