+++ Ghidra/Features/Decompiler/src/decompile/src/bridge.rs
@@ -25,103 +25,20 @@ unsafe impl ExternType for ffi::OpCode {
     type Kind = cxx::kind::Trivial;
 }
 
-#[cxx::bridge]
-pub(crate) mod ffi {
-    
-    extern "Rust" {
-        type Patches;
-        unsafe fn new_patches(arch: *mut Architecture) -> Box<Patches>;
-        fn add_patch(self: &mut Patches, space: &CxxString, offset: usize, size: i32, payload: &CxxString);
-        unsafe fn resolve_patch(self: &Patches, addr: &Address, emit: *mut PcodeEmit) -> i32;
-    }
+// UGLY ALERT! These complex duplication is due to the lack of supporting
+// debug asserton (check for debug build) in #[cxx::bridge].
+// This should be resolved once cxx bridge support such conditional compilation.
+#[cfg(debug_assertions)]
+pub mod debug;
 
-    unsafe extern "C++" {
-        include!("fspec.hh");
-        include!("varnode.hh");
-        include!("pcoderaw.hh");
-        include!("architecture.hh");
-        include!("space.hh");
-        include!("address.hh");
-        include!("translate.hh");
-        include!("libdecomp.hh");
-        include!("interface.hh");
-        include!("consolemain.hh");
-        include!("ifacedecomp.hh");
-        include!("ruststream.hh");
-        include!("ghidra_process.hh");
+#[cfg(debug_assertions)]
+pub use self::debug as ffi;
 
-        type OpCode = crate::model::OpCode;
-        type Address;
-        type AddrSpace;
-        type VarnodeData;
-        type AddrSpaceManager;
-        type Architecture;
-        type PcodeEmit;
-        type IfaceStatus;
-        type IfaceData;
-        type IfaceCommand;
-        type StreamReader;
+#[cfg(not(debug_assertions))]
+pub mod release;
 
-        fn ghidra_process_main();
-
-        fn getName(self: &AddrSpace) -> &CxxString;
-
-        unsafe fn new_address(space: *mut AddrSpace, off: usize) -> UniquePtr<Address>;
-        fn getSpace(self: &Address) -> *mut AddrSpace;
-        fn getOffset(self: &Address) -> usize;
-
-        unsafe fn new_varnode_data(
-            space: *mut AddrSpace,
-            offset: usize,
-            size: u32,
-        ) -> UniquePtr<VarnodeData>;
-
-        fn getAddrSpaceManager(self: &Architecture) -> &AddrSpaceManager;
-
-        fn getSpaceByName(self: &AddrSpaceManager, name: &CxxString) -> *mut AddrSpace;
-
-        unsafe fn dump_rust(
-            emit: *mut PcodeEmit,
-            addr: &Address,
-            opcode: OpCode,
-            out_var: UniquePtr<VarnodeData>,
-            input_vars: &[UniquePtr<VarnodeData>],
-            size: i32,
-        );
-
-        fn startDecompilerLibrary(sleigh_home: &str);
-
-        fn new_iface_status_stub() -> UniquePtr<IfaceStatus>;
-
-
-        unsafe fn call_cmd(cmd: Pin<&mut IfaceCommand>, s: &str);
-        fn getModuleRust(self: &IfaceCommand) -> String;
-        fn createData(self: Pin<&mut IfaceCommand>) -> *mut IfaceData;
-        unsafe fn setData(
-            self: Pin<&mut IfaceCommand>,
-            root: *mut IfaceStatus,
-            data: *mut IfaceData,
-        );
-
-        fn new_load_file_command() -> UniquePtr<IfaceCommand>;
-        fn new_add_path_command() -> UniquePtr<IfaceCommand>;
-        fn new_save_command() -> UniquePtr<IfaceCommand>;
-        fn new_restore_command() -> UniquePtr<IfaceCommand>;
-
-        fn console_main_rust(args: &[String]) -> i32;
-
-        fn new_decompile_command() -> UniquePtr<IfaceCommand>;
-        fn new_print_raw_command() -> UniquePtr<IfaceCommand>;
-        fn new_print_c_command() -> UniquePtr<IfaceCommand>;
-        fn new_addressrange_load_command() -> UniquePtr<IfaceCommand>;
-        fn new_load_func_command() -> UniquePtr<IfaceCommand>;
-
-        fn read(self: Pin<&mut StreamReader>, buf: &mut [u8]) -> usize;
-
-        // opcode
-        fn get_opcode(s: &CxxString) -> OpCode;
-    }
-}
+#[cfg(not(debug_assertions))]
+pub use self::release as ffi;
 
 struct RustReader<'a> {
     reader: Pin<&'a mut ffi::StreamReader>,
